Refactor domain transfer models
