+++ epp-client/src/domain.rs
@@ -3,3 +3,4 @@ pub mod create;
 pub mod delete;
 pub mod info;
 pub mod renew;
+pub mod transfer;
