+++ epp-client/src/epp.rs
@@ -12,7 +12,6 @@ pub use request::contact::info::*;
 pub use request::contact::update::*;
 pub use request::domain::rgp::report::*;
 pub use request::domain::rgp::request::*;
-pub use request::domain::transfer::*;
 pub use request::domain::update::*;
 pub use request::host::check::*;
 pub use request::host::create::*;
@@ -29,7 +28,6 @@ pub use response::contact::info::*;
 pub use response::contact::update::*;
 pub use response::domain::rgp::report::*;
 pub use response::domain::rgp::request::*;
-pub use response::domain::transfer::*;
 pub use response::domain::update::*;
 pub use response::host::check::*;
 pub use response::host::create::*;
