+++ epp-client/src/epp/request/domain.rs
@@ -1,5 +1,4 @@
 //! Types for EPP domain requests
 
 pub mod rgp;
-pub mod transfer;
 pub mod update;
