+++ epp-client/src/epp/response/domain.rs
@@ -1,5 +1,4 @@
 //! Types for EPP domain responses
 
 pub mod rgp;
-pub mod transfer;
 pub mod update;
